# swagger_ui/retrieve/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Definition of ArmRay backend API for data retrieval operations"
  version: "0.0.1"
  title: "ArmRay API"
host: "ec2-52-36-38-196.us-west-2.compute.amazonaws.com:1151"
basePath: "/armray/api/0.1/rest/retrieve"
schemes:
- "http"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /User/{id}:
    get:
      tags:
      - "User"
      description: "Retrieve a 'User' by their id.\n"
      operationId: "getUserById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "the `id` of the `User` to delete"
        required: true
        type: "string"
      responses:
        200:
          description: "The `id` of the deleted `User`"
          schema:
            type: "string"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "User"
  /User/{id}/city:
    get:
      tags:
      - "User"
      description: "Retrieve city associated with a `User`.\n"
      operationId: "getCityNameById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `id` of the `User` to retrieve"
        required: true
        type: "string"
      responses:
        200:
          description: "The city"
          schema:
            type: "string"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "User"
  /User/{id}/city/{city}:
    get:
      tags:
      - "User"
      description: "Check if city is on `User`.\n"
      operationId: "isUserCity"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `id` of the `User` to retrieve"
        required: true
        type: "string"
      - name: "city"
        in: "path"
        description: "The city"
        required: true
        type: "string"
      responses:
        200:
          description: "The boolean"
          schema:
            type: "boolean"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "User"
  /User/{id}/collaborator:
    get:
      tags:
      - "User"
      description: "Retrieve if `User` is a collaborator.\n"
      operationId: "isCollaboratorById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `id` of the `User` to retrieve"
        required: true
        type: "string"
      responses:
        200:
          description: "An array of keywords"
          schema:
            type: "boolean"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "User"
...
  /User/{id}/creator:
    get:
      tags:
      - "User"
      description: "Retrieve if `User` is a creator.\n"
      operationId: "isCreatorById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `id` of the `User` to retrieve"
        required: true
        type: "string"
      responses:
        200:
          description: "An array of keywords"
          schema:
            type: "boolean"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "User"
  /User/{id}/first_name:
    get:
      tags:
      - "User"
      description: "Retrieve first name associated with a `User`.\n"
      operationId: "getFirstNameById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `id` of the `User` to retrieve"
        required: true
        type: "string"
      responses:
        200:
          description: "An array of keywords"
          schema:
            type: "array"
            items:
              type: "string"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "User"
  /User/{id}/first_name/{first_name}:
    get:
      tags:
      - "User"
      description: "Check if first name is on `User`\n"
      operationId: "isUserFirstName"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `id` of the `User` to retrieve"
        required: true
        type: "string"
      - name: "first_name"
        in: "path"
        description: "The keywords"
        required: true
        type: "string"
      responses:
        200:
          description: "The boolean"
          schema:
            type: "boolean"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "User"
  /User/{id}/images:
    get:
      tags:
      - "User"
      - "TODO-create-images-object"
      description: "Retrieve all images associated with a `User`.\n"
      operationId: "getAllUserImages"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `id` of the `User` to retrieve"
        required: true
        type: "string"
      responses:
        200:
          description: "The image url associated with the `User`"
          schema:
            type: "string"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "User"
  /User/{id}/keys:
    get:
      tags:
      - "User"
      description: "Retrieve all keywords associated with a `User`.\n"
      operationId: "getAllUserKeywords"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `id` of the `User` to retrieve"
        required: true
        type: "string"
      responses:
        200:
          description: "An array of keywords"
          schema:
            type: "array"
            items:
              type: "string"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "User"
  /User/{id}/keys/{keyword}:
    get:
      tags:
      - "User"
      description: "Check if a keyword is on a `User`.\n"
      operationId: "isUserHaveKeyword"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `id` of the `User` to retrieve"
        required: true
        type: "string"
      - name: "keyword"
        in: "path"
        description: "The keyword"
        required: true
        type: "string"
      responses:
        200:
          description: "The boolean"
          schema:
            type: "boolean"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "User"
  /User/{id}/last_name:
    get:
      tags:
      - "User"
      description: "Retrieve first name associated with a `User`.\n"
      operationId: "getLastNameById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `id` of the `User` to retrieve"
        required: true
        type: "string"
      responses:
        200:
          description: "An array of keywords"
          schema:
            type: "array"
            items:
              type: "string"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "User"
  /User/{id}/last_name/{last_name}:
    get:
      tags:
      - "User"
      description: "Check if last name is on `User`.\n"
      operationId: "isUserLastName"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `id` of the `User` to retrieve"
        required: true
        type: "string"
      - name: "last_name"
        in: "path"
        description: "The keywords"
        required: true
        type: "string"
      responses:
        200:
          description: "The boolean"
          schema:
            type: "boolean"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "User"
...
  /User/{id}/project:
    get:
      tags:
      - "User"
      - "Project"
      description: "Retrieve all 'Project' associated with a `User`.\n"
      operationId: "getAllUserProjectById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `id` of the `User` to retrieve"
        required: true
        type: "string"
      responses:
        200:
          description: "The `Project` associated with the `User`"
          schema:
            $ref: "#/definitions/Project"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "User"
  /User/{id}/project/{project_id}:
    get:
      tags:
      - "User"
      - "Project"
      description: "Is 'Project' associated with a `User`?\n"
      operationId: "isUserOnProjectById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `id` of the `User` to retrieve"
        required: true
        type: "string"
      - name: "project_id"
        in: "path"
        description: "The `project_id` of the `Project` to retrieve"
        required: true
        type: "string"
      responses:
        200:
          description: "The boolean"
          schema:
            type: "boolean"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "User"
  /User/{id}/review:
    get:
      tags:
      - "User"
      - "Review"
      description: "Retrieve all 'Review' associated with a `User`.\n"
      operationId: "getAllUserReviewById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `id` of the `User` to retrieve"
        required: true
        type: "string"
      responses:
        200:
          description: "An array of reviews"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Review"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "User"
  /User/{id}/review/{review_id}:
    get:
      tags:
      - "User"
      - "Review"
      description: "Is a 'Review' associated with a `User`??\n"
      operationId: "isReviewOnUserById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `id` of the `User` to retrieve"
        required: true
        type: "string"
      - name: "review_id"
        in: "path"
        description: "The `review_id` of the `Project` to retrieve"
        required: true
        type: "string"
      responses:
        200:
          description: "The boolean"
          schema:
            type: "boolean"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "User"
  /User/{id}/skills:
    get:
      tags:
      - "User"
      - "Skill"
      description: "Retrieve all skills associated with a `User`.\n"
      operationId: "getAllUserSkills"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `id` of the `User` to retrieve"
        required: true
        type: "string"
      responses:
        200:
          description: "An array of UserSkills ids"
          schema:
            type: "array"
            items:
              type: "string"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "User"
  /User/{id}/skills/{skill_id}:
    get:
      tags:
      - "User"
      - "Skill"
      description: "Check if `Skill`is associated with a `User`.\n"
      operationId: "getIsSkillOnUser"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `id` of the `User` to retrieve"
        required: true
        type: "string"
      - name: "skill_id"
        in: "path"
        description: "The `id` of the `Skill` to retrieve"
        required: true
        type: "string"
      responses:
        200:
          description: "An array of UserSkills ids"
          schema:
            type: "array"
            items:
              type: "string"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "User"
  /User/{id}/state:
    get:
      tags:
      - "User"
      description: "Retrieve state associated with a `User`.\n"
      operationId: "getStateById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `id` of the `User` to retrieve"
        required: true
        type: "string"
      responses:
        200:
          description: "The state"
          schema:
            type: "string"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "User"
  /User/{id}/state/{state}:
    get:
      tags:
      - "User"
      description: "Check if state is on `User`.\n"
      operationId: "isUserState"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `id` of the `User` to retrieve"
        required: true
        type: "string"
      - name: "state"
        in: "path"
        description: "The state"
        required: true
        type: "string"
      responses:
        200:
          description: "The boolean"
          schema:
            type: "boolean"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "User"
  /User/{id}/uni/{university_id}:
    get:
      tags:
      - "User"
      - "University"
      description: "Is 'University' associated with a `User`?\n"
      operationId: "isUniversityOnUser"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `id` of the `User` to retrieve"
        required: true
        type: "string"
      - name: "university_id"
        in: "path"
        description: "The `university_id` of the `University` to retrieve"
        required: true
        type: "string"
      responses:
        200:
          description: "The `University` associated with the `User`"
          schema:
            type: "boolean"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "User"
...
  /UserSkill/{id}/user:
    get:
      tags:
      - "Skill"
      - "UserSkill"
      - "User"
      description: "Gets the `user`\n"
      operationId: "getUserSkillUserById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `Skill` id"
        required: true
        type: "string"
      responses:
        200:
          description: "The `Skill`"
          schema:
            $ref: "#/definitions/Skill"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Skill"
...
definitions:
  Error:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      fields:
        type: "string"
      message:
        type: "string"
...
  Project:
    type: "object"
    properties:
      created_date:
        type: "string"
        format: "date-time"
        description: "datetime stamp of `Project`creation"
      grant_facilitator:
        type: "string"
        description: "Name of individual fascilitating `Grant` funding"
      grant_ids:
        type: "array"
        description: "Array of `Grant` IDs associated with the `Project`"
        items:
          type: "string"
      is_project_private:
        type: "boolean"
        description: "is `Project`private?"
      posts:
        type: "array"
        description: "posts related to project"
        items:
          type: "string"
      project_collaborators_id:
        type: "array"
        description: "array of `Project`collaborator ids"
        items:
          type: "string"
      project_creator_id:
        type: "array"
        description: "array of `Project`creator ids"
        items:
          type: "string"
      project_description:
        type: "string"
        description: "Description of project"
      project_file:
        type: "array"
        description: "Paths to files for `Grant`"
        items:
          type: "string"
      project_id:
        type: "string"
        description: "Unique identifier representing a specific project"
      project_keywords:
        type: "array"
        description: "array of `Project`keywords"
        items:
          type: "string"
      project_name:
        type: "string"
        description: "Project's name"
      project_urls:
        type: "array"
        description: "array of `Project`URLs"
        items:
          type: "string"
      reviews:
        type: "array"
        description: "reviews"
        items:
          type: "string"
      university_id:
        type: "string"
        description: "ID of `University`hosting research"
      updated_date:
        type: "string"
        format: "date-time"
        description: "datetime stamp of `Project`update"
...
  Review:
    type: "object"
    properties:
      created_date:
        type: "string"
        format: "date-time"
        description: "datetime stamp of `Review`creation"
      rating:
        type: "integer"
        format: "int32"
        description: "rating, must be 1-5"
      review_body:
        type: "string"
        description: "text body of review"
      review_id:
        type: "string"
        description: "id of review"
      reviewer:
        description: "reviewer user"
        $ref: "#/definitions/User"
      reviewer_id:
        type: "string"
        description: "reviewer `User`id"
      updated_date:
        type: "string"
        format: "date-time"
        description: "datetime stamp of `Review`update"
...